requirements.txt
anyio==4.12.1
asttokens==3.0.1
bottle==0.13.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clr_loader==0.2.10
colorama==0.4.6
comm==0.2.3
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.8.19
decorator==5.2.1
diskcache==5.6.3
executing==2.2.1
filelock==3.20.2
flet==0.80.1
flet-desktop==0.80.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
llama_cpp_python==0.3.16
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
proxy_tools==0.1.0
psutil==7.2.1
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
PyPDF2==3.0.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pythonnet==3.0.5
pywebview==6.1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
repath==0.9.0
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing==3.7.4.3
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.2.14
